ai-coin-bot/requirements.txt
# AI Coin Buyback Bot Dependencies
# =================================

# OpenAI API
openai>=1.12.0

# Solana
solana>=0.32.0
solders>=0.20.0
base58>=2.1.1

# HTTP Client
httpx>=0.26.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Data Processing
numpy>=1.26.0

# Utilities
rich>=13.7.0  # Beautiful terminal output (optional)
